+++ menu/src/main/scala/de/welt/contentapi/menu/services/AdminMenuService.scala
@@ -2,37 +2,42 @@ package de.welt.contentapi.menu.services
 
 import com.google.inject.ImplementedBy
 import de.welt.contentapi.core.client.services.s3.S3Client
-import de.welt.contentapi.menu.models.{Menu, MenuMetadata}
+import de.welt.contentapi.menu.models.{ApiMenu, ApiMenuMetadata}
 import javax.inject.{Inject, Singleton}
 import play.api.libs.json.Json
 import play.api.{Configuration, Environment, Logger}
 
 @ImplementedBy(classOf[AdminMenuServiceImpl])
-trait AdminMenuService {
-  def get(): Option[Menu]
-  def save(menuData: Menu, user: String): Menu
+trait AdminMenuService extends MenuService {
+  def save(menuData: ApiMenu, user: String): ApiMenu
 }
 
 @Singleton
 class AdminMenuServiceImpl @Inject()(config: Configuration,
                                      environment: Environment,
-                                     s3: S3Client) extends AdminMenuService with MenuServiceUtils {
+                                     s3: S3Client) extends MenuServiceRepository(config, environment.mode) with AdminMenuService {
 
   import de.welt.contentapi.menu.models.MenuFormats._
 
-  private[services] val s3Config: S3Config = s3Config(config, environment.mode)
+  override def get(): ApiMenu = loadMenu(s3)
+    .getOrElse(throw new IllegalStateException(s"There was an error to load the menu json file from S3. Check if the config is correct: $s3Config"))
 
-  override def get(): Option[Menu] = loadMenu(s3, s3Config)
+  override def save(menuData: ApiMenu, user: String): ApiMenu = {
+    if (menuData.isEmpty) {
+      throw new IllegalStateException(
+        """
+          |It is not allowed to save a empty menu. Please check your input data.
+          |When this is not an error fix the json file on S3 by hand. Sorry.
+        """.stripMargin)
+    } else {
+      val menuDataWithUpdatedMetadata: ApiMenu = menuData.copy(metadata = ApiMenuMetadata(changedBy = user))
+      val serializedMenuData = Json.toJson(menuDataWithUpdatedMetadata).toString()
 
-  override def save(menuData: Menu, user: String): Menu = {
-    val menuDataWithUpdatedMetadata: Menu = menuData.copy(metadata = MenuMetadata(changedBy = user))
-    val serializedMenuData = Json.toJson(menuDataWithUpdatedMetadata).toString()
-    val filePath: String = s3FilePath(s3Config)
+      Logger.info(s"saving menu data to s3 -> ${s3Config.fullFilePath} - user: $user")
 
-    Logger.info(s"saving menu data to s3 -> $s3Config.bucket/$filePath - user: $user")
+      s3.putPrivate(s3Config.bucket, s3Config.fullFilePath, serializedMenuData, "application/json")
 
-    s3.putPrivate(s3Config.bucket, filePath, serializedMenuData, "application/json")
-
-    menuDataWithUpdatedMetadata
+      menuDataWithUpdatedMetadata
+    }
   }
 }