+++ menu/src/main/scala/de/welt/contentapi/menu/services/MenuService.scala
@@ -5,7 +5,7 @@ import java.time.Instant
 import com.google.inject.ImplementedBy
 import de.welt.contentapi.core.client.services.CapiExecutionContext
 import de.welt.contentapi.core.client.services.s3.S3Client
-import de.welt.contentapi.menu.models.Menu
+import de.welt.contentapi.menu.models.ApiMenu
 import javax.inject.{Inject, Singleton}
 import play.api.{Configuration, Environment, Logger, Mode}
 
@@ -13,18 +13,16 @@ import scala.concurrent.duration._
 
 @ImplementedBy(classOf[MenuServiceImpl])
 trait MenuService {
-  def get(): Menu
+  def get(): ApiMenu
 }
 
 @Singleton
 class MenuServiceImpl @Inject()(config: Configuration,
                                 s3: S3Client,
                                 environment: Environment,
-                                implicit val capi: CapiExecutionContext) extends MenuService with MenuServiceUtils {
+                                implicit val capi: CapiExecutionContext) extends MenuServiceRepository(config, environment.mode) with MenuService {
 
-  private[services] val s3Config: S3Config = s3Config(config, environment.mode)
-
-  private[services] var menu: Menu = Menu()
+  private[services] var menu: ApiMenu = ApiMenu()
   private[services] var lastModified: Instant = Instant.MIN
 
   if (Mode.Test != environment.mode) {
@@ -34,23 +32,23 @@ class MenuServiceImpl @Inject()(config: Configuration,
     Logger.info("Menu Data will not be loaded when started in Mode.Test. Please mock it.")
   }
 
-  override def get(): Menu = menu
+  override def get(): ApiMenu = menu
 
   private def syncData() = {
-    val maybeLastMod: Option[Instant] = s3.getLastModified(s3Config.bucket, s3FilePath(s3Config))
-    Logger.info(s"Starting MenuService.syncData(), lastmod= $maybeLastMod")
+    val maybeLastMod: Option[Instant] = s3.getLastModified(s3Config.bucket, s3Config.fullFilePath)
+    Logger.info(s"Starting MenuService.syncData(), LastMod=$maybeLastMod")
 
     for {
       remoteLastModified: Instant ← maybeLastMod
       if lastModified != remoteLastModified
     } yield {
-      val maybeFreshMenu: Option[Menu] = loadMenu(s3, s3Config)
+      val maybeFreshMenu: Option[ApiMenu] = loadMenu(s3)
 
       maybeFreshMenu.foreach(freshMenu ⇒ {
         menu = freshMenu
         lastModified = remoteLastModified
 
-        Logger.info(s"Finished MenuService.syncData(). LastMod $lastModified -> $remoteLastModified")
+        Logger.info(s"Finished MenuService.syncData(). LastMod=$lastModified -> $remoteLastModified")
       })
     }
   }